Refactor: Rename BookClub -> blogs
Renames the project settings directory
Refactors changes in settings.py
Added .pylintrc to gitignore

Delete __init__.py

Swap names
