+++ .gitignore
@@ -6,6 +6,7 @@
 *.pot
 *.pyc
 __pycache__/
+.pylintrc
 local_settings.py
 db.sqlite3
 db.sqlite3-journal
