+++ bookclub/urls.py
@@ -15,7 +15,7 @@ Including another URLconf
 """
 from django.contrib import admin
 from django.urls import path
-from bookclub import views
+from blogs import views
 
 urlpatterns = [
     path('admin/', admin.site.urls),
